refactor: use password_validation module
Only for changing password so; may need to update StrongPasswordAuthenticationForm to match
