+++ ietf/ietfauth/validators.py
@@ -4,7 +4,6 @@ import re
 from django import forms
 from django.conf import settings
 from django.core.exceptions import ValidationError
-from zxcvbn import zxcvbn
 
 
 def prevent_at_symbol(name):
@@ -33,23 +32,3 @@ def is_allowed_address(value):
             raise ValidationError(
                 "This email address is not valid in a datatracker account"
             )
-
-
-class StrongPasswordValidator:
-    message = "This password does not meet complexity requirements and is easily guessable."
-    code = "weak"
-    min_zxcvbn_score = 3
-
-    def __init__(self, message=None, code=None, min_zxcvbn_score=None):
-        if message is not None:
-            self.message = message
-        if code is not None:
-            self.code = code
-        if min_zxcvbn_score is not None:
-            self.min_zxcvbn_score = min_zxcvbn_score
-
-    def __call__(self, password):
-        """Validate that a password is strong enough"""
-        strength_report = zxcvbn(password[:72], max_length=72)
-        if strength_report["score"] < self.min_zxcvbn_score:
-            raise ValidationError(message=self.message, code=self.code)
